refactor: re-home navigator sim renderers
Move the navigator-side simulated renderers under nav_model/, stating
package ownership: nav_model/sim_body.py is the predicted smooth
Lambert ellipsoid (crater texture is truth-side and lives only in
sim/forward/body.py) and nav_model/sim_ring.py is the predicted
annulus/gap template.  Both build on the shared sim geometry modules,
so the image side and the prediction side keep identical conventions.

The crater-shading test moves to the image-side renderer and pins a
new crater seed: the Generator reseed changed realizations, and the
chosen seed keeps placement asymmetry (~0.15 px) far below the 0.5 px
tolerance that guards the ~9 px shading-frame defect.
