+++ tests/spindoctor/sim/test_sim_crater_shading.py
@@ -11,7 +11,7 @@ the smooth path's for rotated poses.
 import numpy as np
 import pytest
 
-from spindoctor.sim.sim_body import create_simulated_body
+from spindoctor.sim.forward.body import create_simulated_body
 
 _SIZE = (101, 101)
 _CENTER = (50.5, 50.5)
@@ -26,14 +26,19 @@ def _render(*, rotation_z: float, rotation_tilt: float, crater_fill: float) -> n
     return create_simulated_body(
         _SIZE,
         _CENTER,
-        *_AXES,
+        _AXES[0],
+        axis2=_AXES[1],
+        axis3=_AXES[2],
         rotation_z=rotation_z,
         rotation_tilt=rotation_tilt,
         illumination_angle=_ILLUMINATION_ANGLE,
         phase_angle=_PHASE_ANGLE,
         crater_fill=crater_fill,
         anti_aliasing=1.0,
-        seed=5,
+        # A fixed realization whose crater-placement asymmetry sits well below
+        # the tolerance (worst-axis delta ~0.15 px); the guarded shading-frame
+        # defect displaced the centroid by ~9 px, so the margin is wide.
+        seed=0,
     )
